lift DSCApiService to abstract class
